requirements.txt
absl-py==2.3.0
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
alpaca-py==0.35.0
alpaca-trade-api==3.2.0
alpha_vantage==3.0.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
appdirs==1.4.4
appnope==0.1.4
APScheduler==3.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==24.3.0
babel==2.17.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.1
ccxt==4.4.45
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cmdstanpy==1.2.5
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
duckdb==1.1.3
exchange_calendars==4.7
executing==2.1.0
fastapi==0.115.7
fastjsonschema==2.21.1
filelock==3.16.1
finnhub-python==2.4.22
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.55.3
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
groq==0.13.1
grpcio==1.73.0
h11==0.14.0
h5py==3.14.0
holidays==0.63
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
ibapi==9.81.1.post1
idna==3.10
ijson==3.3.0
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.31.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json5==0.12.0
jsonpickle==4.0.1
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.10.0
kiwisolver==1.4.8
korean-lunar-calendar==0.3.1
libclang==18.1.1
lumibot==3.8.24
lumiwealth-tradier==0.1.14
lxml==5.3.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml-dtypes==0.3.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.0.3
multidict==6.1.0
multitasking==0.0.11
namex==0.1.0
narwhals==1.23.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
newsapi-python==0.2.7
nltk==3.9.1
notebook==7.4.2
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
openai==1.58.1
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandas_market_calendars==4.5.1
pandocfilters==1.5.1
parso==0.8.4
peewee==3.17.8
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
polygon-api-client==1.14.2
praw==7.8.1
prawcore==2.4.0
prometheus_client==0.22.0
prompt_toolkit==3.0.48
propcache==0.2.1
prophet==1.1.6
protobuf==4.25.8
psutil==6.1.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pycares==4.5.0
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
pyluach==2.2.0
pyparsing==3.2.1
PyPDF2==3.0.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
pytz==2024.2
PyYAML==6.0.1
pyzmq==26.4.0
Quandl==3.7.0
quantstats-lumi==0.3.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.0
scikit-learn==1.6.0
scipy==1.15.0
seaborn==0.13.2
selenium==4.29.0
Send2Trash==1.8.3
setuptools==80.9.0
sip==6.12.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
sseclient-py==1.8.0
stack-data==0.6.3
stanio==0.5.1
starlette==0.45.3
streamlit==1.41.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-macos==2.16.2
termcolor==2.5.0
terminado==0.18.1
textblob==0.18.0.post0
thetadata==0.9.11
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
toolz==1.0.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
trio==0.29.0
trio-websocket==0.12.1
tweepy==4.14.0
types-python-dateutil==2.9.0.20250516
typing_extensions==4.5.0
tzdata==2024.2
tzlocal==5.2
update-checker==0.18.0
uri-template==1.3.0
urllib3==1.26.20
vaderSentiment==3.3.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
wget==3.2
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
yfinance==0.2.51
